requirements_images.txt
# 画像生成とPDF作成に必要な依存関係
requests>=2.25.1
Pillow>=8.0.0
pathlib2>=2.3.5; python_version < "3.4" 